Month1-6/Month4_NLP_CV_LangChain/TimeSeries/Week16_TimeSeries/Project/requirements.txt
treamlit
plotly 
prophet 
statsmodels 
pandas
